requirements.txt
Flask==3.0.0
flask-cors==4.0.0
torch==2.1.0
transformers==4.37.0
Pillow==10.2.0
psutil==5.9.6
accelerate==0.25.0
sentencepiece==0.1.99
protobuf==4.25.1
safetensors==0.4.1
gunicorn==21.2.0
python-dotenv==1.0.0


